% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_volclustR_app.R
\name{run_volclustR_app}
\alias{run_volclustR_app}
\title{Launch the VolClusR Interactive Dashboard}
\usage{
run_volclustR_app()
}
\description{
Starts a professional, local Shiny web application to interactively
upload OHLCV data, execute the C++ engine, and visualize EM clustering.
}
